requirements.txt
fastapi
uvicorn
google-genai
websockets
python-dotenv
python-multipart
pypdf
python-docx
httpx
beautifulsoup4
passlib[bcrypt]
python-jose[cryptography]
asyncpg
pytest
